@@ -142,7 +142,7 @@ def agents_page(request: Request):
                 "slug": agent.slug,
                 "avatar": agent.avatar,
                 "name": agent.name,
-                "tuning": agent.tuning,
+                "personality": agent.personality,
                 "public": agent.public,
                 "creator": agent.creator.username if agent.creator else None,
                 "managed_by_admin": agent.managed_by_admin,
@@ -186,7 +186,7 @@ def agent_page(request: Request, agent_slug: str):
         "slug": agent.slug,
         "avatar": agent.avatar,
         "name": agent.name,
-        "tuning": agent.tuning,
+        "personality": agent.personality,
         "public": agent.public,
         "creator": agent.creator.username if agent.creator else None,
         "managed_by_admin": agent.managed_by_admin,
