@@ -70,15 +70,12 @@ class SearchModels:
 
 
 @dataclass
-class GPT4AllProcessorConfig:
+class OfflineChatProcessorConfig:
     loaded_model: Union[Any, None] = None
 
 
-class GPT4AllProcessorModel:
-    def __init__(
-        self,
-        chat_model: str = "mistral-7b-instruct-v0.1.Q4_0.gguf",
-    ):
+class OfflineChatProcessorModel:
+    def __init__(self, chat_model: str = "NousResearch/Hermes-2-Pro-Mistral-7B-GGUF"):
         self.chat_model = chat_model
         self.loaded_model = None
         try:
